[GPU Sim] Triangle meshes, debug features, updated API for ModelParser, OptiX optimisations, better CLI support for TestSimulator

+++ GPUSim/MolTracer.h
@@ -16,7 +16,7 @@ namespace flowgpu {
 
         void launchRT();
 
-        unsigned int fetchData();
+        unsigned long long int fetchData();
 
         void run();
